Added a third Neighbor type: "Empty"
I added a third neighbor in the class and the code to manage it, to know if there is some other Tile instead just "not this tile", so you can make a grid with a lot of tiles but can know at least too when the "map reach the end" to show the right Tile. Added the needed code in RuleTilesEditor.cs too.