Added a third Neighbor: "Empty"
I added a third neighbor in the class and the code to manage it, to know if there is some other Tile instead just "not this tile", so you can make a grid with a lot of tiles but can know at least too when the "map reach the end" to show the right Tile. The new neighbor include their own base64 texture. Added the needed code in RuleTiles.cs too.

Based on an fix of this same problem in a old version from: https://gist.github.com/Bahamutho/aa2d2f12f60c71eaf462d8126505e54a